Commit4
Added the main panel which display the main components of the app interface.
An ImagePanel that contains a resizable image.
A RoundMenuBar that reprezent a menu bar with round borders.
Added the action listener to the login button from login data panel and made the connection with the main panel